Split bvh module into multiple files
